[jira] Verify that projects referenced by issues exist (#34)
* Handle multiple issues in single message

This provides proper handling of multiple issues in single message. We
use different register command so we can reply with multiple replies

* Query JIRA with basic auth to verify projects exist

With this change the bot will not respond to bogus mentions of issues
if their projects do not actually exist in configured JIRA instance.
This check is done using basic auth and JIRA api query.

This needs 3 additional env variables to be set up:
 * JIRA_BASE_URL
 * JIRA_USER
 * JIRA_PASS

* Add JIRA plugin readme file

* Remove original env variable and update main README

We point in jira specific readme for configuration information. Also got
rid of original env variable - the URL is now automatically created from
base URL env variable